Update clear screen tests to run on all supported OSes
