+++ .github/workflows/pullRequests.yaml
@@ -20,6 +20,7 @@ jobs:
     strategy:
       matrix:
         dir: [".", "example"]
+        os: [ubuntu-latest, windows-latest, macos-latest]
     runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v4
