+++ clearScreen_windows_test.go
@@ -1,0 +1,10 @@
+//go:build windows
+// +build windows
+
+package wmenu
+
+import "testing"
+
+func TestClearWindows(t *testing.T) {
+	clearOs("windows")
+}
